change type of vars to fixed bit lenght types
git-svn-id: https://svn.mi.fu-berlin.de/des-testbed/Software/DES-SERT/des-aodv/trunk@8142 ec5dfc5b-5239-0410-a031-e37cc464ff32

+++ src/aodv.c
@@ -31,9 +31,9 @@ For further information and questions please use the web site
 #include "pipeline/aodv_pipeline.h"
 #include "database/aodv_database.h"
 
-int hello_size = HELLO_SIZE;
-int hello_interval = HELLO_INTERVAL;
-int rreq_size = RREQ_SIZE;
+uint16_t hello_size = HELLO_SIZE;
+uint16_t hello_interval = HELLO_INTERVAL;
+uint16_t rreq_size = RREQ_SIZE;
 double gossipp = GOSSIPP;
 int dest_only = DESTONLY;
